+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Remove the target version dropdown menu, it will now use the server dropdown menu
 * The server must have a valid status before using the plugin
 * The metadata for each server is stored for maximum two days
+* Fix the shortname generation when the layer name contains a `-`
 
 ## 3.11.5 - 2023-04-11